Add Default Mock Implementation
Add mocks for common Storybook endpoints to use in CI testing. Using those mocks in the tests here.
