+++ .storybook/__conf__/mochaMockConfig.js
@@ -1,14 +1,8 @@
-import {storiesOf, action, linkTo, specs, describe, it,} from "../__mocks__/facade-mocha";
+import {storiesOf, action, linkTo, specs} from "../__mocks__/facade-mocha";
 global.storiesOf = storiesOf;
 global.action = action;
 global.linkTo = linkTo;
 global.specs = specs;
-global.describe = describe;
-global.it = it;
-global.after = after;
-global.before = before;
-global.beforeEach = beforeEach;
-global.afterEach = afterEach;
 
 import { jsdom } from 'jsdom';
